backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
feedparser==6.0.11
openai==1.54.0
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.10.4
pydantic==2.9.2
